fix(security): satisfy phpmd/CI for #314
Split oversized test classes that exceeded phpmd TooManyMethods (max 25)
into cohesive sibling classes sharing abstract base TestCases. No coverage
lost — every test and assertion preserved.
