+++ tests/Unit/User/Application/CommandHandler/CompleteTwoFactorCommandHandlerTest.php
@@ -4,52 +4,11 @@ declare(strict_types=1);
 
 namespace App\Tests\Unit\User\Application\CommandHandler;
 
-use App\Shared\Infrastructure\Factory\UuidFactory as SharedUuidFactory;
-use App\Shared\Infrastructure\Transformer\UuidTransformer;
-use App\Tests\Unit\UnitTestCase;
 use App\User\Application\Command\CompleteTwoFactorCommand;
-use App\User\Application\CommandHandler\CompleteTwoFactorCommandHandler;
-use App\User\Application\DTO\CompleteTwoFactorCommandResponse;
-use App\User\Application\DTO\IssuedSession;
-use App\User\Application\Factory\IssuedSessionFactoryInterface;
-use App\User\Application\Validator\TwoFactorCodeValidatorInterface;
-use App\User\Domain\Entity\PendingTwoFactor;
-use App\User\Domain\Entity\User;
-use App\User\Domain\Factory\UserFactory;
-use App\User\Domain\Repository\PendingTwoFactorRepositoryInterface;
-use App\User\Domain\Repository\UserRepositoryInterface;
-use App\User\Infrastructure\Publisher\TwoFactorPublisherInterface;
-use DateTimeImmutable;
-use PHPUnit\Framework\MockObject\MockObject;
 use Symfony\Component\HttpKernel\Exception\UnauthorizedHttpException;
-use Symfony\Component\Uid\Ulid;
 
-final class CompleteTwoFactorCommandHandlerTest extends UnitTestCase
+final class CompleteTwoFactorCommandHandlerTest extends CompleteTwoFactorCommandHandlerTestCase
 {
-    private UserRepositoryInterface&MockObject $userRepository;
-    private PendingTwoFactorRepositoryInterface&MockObject $pendingTwoFactorRepository;
-    private IssuedSessionFactoryInterface&MockObject $sessionIssuer;
-    private TwoFactorCodeValidatorInterface&MockObject $twoFactorCodeVerifier;
-    private TwoFactorPublisherInterface&MockObject $events;
-    private UserFactory $userFactory;
-    private UuidTransformer $uuidTransformer;
-
-    #[\Override]
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $this->userRepository = $this->createMock(UserRepositoryInterface::class);
-        $this->pendingTwoFactorRepository = $this->createMock(
-            PendingTwoFactorRepositoryInterface::class
-        );
-        $this->sessionIssuer = $this->createMock(IssuedSessionFactoryInterface::class);
-        $this->twoFactorCodeVerifier = $this->createMock(TwoFactorCodeValidatorInterface::class);
-        $this->events = $this->createMock(TwoFactorPublisherInterface::class);
-        $this->userFactory = new UserFactory();
-        $this->uuidTransformer = new UuidTransformer(new SharedUuidFactory());
-    }
-
     public function testInvokeThrowsUnauthorizedWhenPendingSessionIsMissing(): void
     {
         $this->pendingTwoFactorRepository
@@ -164,334 +123,5 @@ final class CompleteTwoFactorCommandHandlerTest extends UnitTestCase
     public function testInvokeRejectsRecoveryCodeWithTrailingCharacter(): void
     {
         $this->assertInvalidTwoFactorCodeRejected('AB1C-D234X');
-    }
-
-    public function testInvokeThrowsUnauthorizedWhenRecoveryCodeVerificationFails(): void
-    {
-        $user = $this->createTwoFactorEnabledUser();
-        $pending = $this->createPendingSession($user->getId(), '+5 minutes');
-        $this->configureLookupsOnce($pending, $user);
-
-        $this->twoFactorCodeVerifier->expects($this->once())
-            ->method('verifyAndResolveMethod')
-            ->with($user, 'EF55-GH66')
-            ->willReturn(null);
-
-        $this->events->expects($this->once())->method('publishFailed');
-        $this->pendingTwoFactorRepository->expects($this->never())->method('consumeIfActive');
-        $this->expectException(UnauthorizedHttpException::class);
-        $this->expectExceptionMessage('Invalid two-factor code.');
-        $this->createHandler()->__invoke(new CompleteTwoFactorCommand(
-            $pending->getId(),
-            'EF55-GH66',
-            $this->faker->ipv4(),
-            $this->faker->userAgent()
-        ));
-    }
-
-    public function testInvokeDoesNotConsumeRecoveryCodeWhenPendingSessionConsumeFails(): void
-    {
-        $user = $this->createTwoFactorEnabledUser();
-        $pending = $this->createPendingSession($user->getId(), '+5 minutes');
-        $this->configureLookupsOnce($pending, $user);
-        $this->twoFactorCodeVerifier->expects($this->once())
-            ->method('verifyAndResolveMethod')
-            ->with($user, 'AB12-CD34')
-            ->willReturn(TwoFactorCodeValidatorInterface::METHOD_RECOVERY_CODE);
-        $this->twoFactorCodeVerifier->expects($this->never())->method('consumeRecoveryCodeOrFail');
-        $this->pendingTwoFactorRepository->expects($this->once())
-            ->method('consumeIfActive')
-            ->with($pending->getId(), $this->isInstanceOf(DateTimeImmutable::class))
-            ->willReturn(false);
-        $this->expectException(UnauthorizedHttpException::class);
-        $this->expectExceptionMessage('Invalid or expired two-factor session.');
-        $this->createHandler()->__invoke($this->createCommand($pending->getId(), 'AB12-CD34'));
-    }
-
-    public function testInvokeThrowsUnauthorizedWhenTotpVerificationFails(): void
-    {
-        $user = $this->createTwoFactorEnabledUser();
-        $pending = $this->createPendingSession($user->getId(), '+5 minutes');
-        $this->configureLookupsOnce($pending, $user);
-
-        $this->twoFactorCodeVerifier->expects($this->once())
-            ->method('verifyAndResolveMethod')
-            ->with($user, '123456')
-            ->willReturn(null);
-
-        $this->events->expects($this->once())->method('publishFailed');
-        $this->pendingTwoFactorRepository->expects($this->never())->method('consumeIfActive');
-        $this->expectException(UnauthorizedHttpException::class);
-        $this->expectExceptionMessage('Invalid two-factor code.');
-        $this->createHandler()->__invoke(new CompleteTwoFactorCommand(
-            $pending->getId(),
-            '123456',
-            $this->faker->ipv4(),
-            $this->faker->userAgent()
-        ));
-    }
-
-    public function testInvokeIncrementsAndPersistsFailedAttemptOnInvalidCode(): void
-    {
-        $user = $this->createTwoFactorEnabledUser();
-        $pending = $this->createPendingSession($user->getId(), '+5 minutes');
-        $this->configureLookupsOnce($pending, $user);
-        $this->twoFactorCodeVerifier->method('verifyAndResolveMethod')->willReturn(null);
-
-        $this->pendingTwoFactorRepository->expects($this->once())
-            ->method('save')
-            ->with($pending);
-        $this->pendingTwoFactorRepository->expects($this->never())->method('delete');
-
-        try {
-            $this->createHandler()->__invoke($this->createCommand($pending->getId(), '123456'));
-            $this->fail('Expected UnauthorizedHttpException.');
-        } catch (UnauthorizedHttpException) {
-            $this->assertSame(1, $pending->getFailedAttempts());
-        }
-    }
-
-    public function testInvokeInvalidatesPendingSessionAfterMaxFailedAttempts(): void
-    {
-        $user = $this->createTwoFactorEnabledUser();
-        $pending = $this->createPendingSession($user->getId(), '+5 minutes');
-        $pending->setFailedAttempts(PendingTwoFactor::MAX_FAILED_ATTEMPTS - 1);
-        $this->configureLookupsOnce($pending, $user);
-        $this->twoFactorCodeVerifier->method('verifyAndResolveMethod')->willReturn(null);
-
-        $this->pendingTwoFactorRepository->expects($this->once())
-            ->method('delete')
-            ->with($pending);
-        $this->pendingTwoFactorRepository->expects($this->never())->method('save');
-
-        $this->expectException(UnauthorizedHttpException::class);
-        $this->expectExceptionMessage('Invalid two-factor code.');
-        $this->createHandler()->__invoke($this->createCommand($pending->getId(), '123456'));
-    }
-
-    public function testInvokeRejectsPendingSessionThatAlreadyExhaustedAttempts(): void
-    {
-        $user = $this->createTwoFactorEnabledUser();
-        $pending = $this->createPendingSession($user->getId(), '+5 minutes');
-        $pending->setFailedAttempts(PendingTwoFactor::MAX_FAILED_ATTEMPTS);
-        $this->pendingTwoFactorRepository->expects($this->once())
-            ->method('findById')
-            ->with($pending->getId())
-            ->willReturn($pending);
-        $this->userRepository->expects($this->never())->method('findById');
-        $this->twoFactorCodeVerifier->expects($this->never())->method('verifyAndResolveMethod');
-
-        $this->expectException(UnauthorizedHttpException::class);
-        $this->expectExceptionMessage('Invalid or expired two-factor session.');
-        $this->createHandler()->__invoke($this->createCommand($pending->getId(), '123456'));
-    }
-
-    public function testInvokeSucceedsWithValidTotpCode(): void
-    {
-        $user = $this->createTwoFactorEnabledUser();
-        $pending = $this->createPendingSession($user->getId(), '+5 minutes');
-        $this->configureLookupsOnce($pending, $user);
-        $this->expectTotpVerification($user, '123456');
-
-        $issued = new IssuedSession('session-id', 'access-token', 'refresh-token');
-        $this->expectIssuedSession($issued);
-        $this->pendingTwoFactorRepository->expects($this->once())
-            ->method('consumeIfActive')
-            ->with($pending->getId(), $this->isInstanceOf(DateTimeImmutable::class))
-            ->willReturn(true);
-        $this->events->expects($this->once())->method('publishCompleted');
-        $command = new CompleteTwoFactorCommand(
-            $pending->getId(),
-            '123456',
-            $this->faker->ipv4(),
-            $this->faker->userAgent()
-        );
-
-        $response = $this->createHandler()->__invoke($command);
-        $this->assertResponseTokens($response, 'access-token', 'refresh-token');
-    }
-
-    public function testInvokeFailsWithoutCountingAttemptWhenRecoveryConsumeFails(): void
-    {
-        $user = $this->createTwoFactorEnabledUser();
-        $pending = $this->createPendingSession($user->getId(), '+5 minutes');
-        $this->arrangeRecoveryConsumeFailureAfterSessionConsumed($pending, $user);
-
-        try {
-            $this->createHandler()->__invoke($this->createCommand($pending->getId(), 'AB12-CD34'));
-            $this->fail('Expected UnauthorizedHttpException.');
-        } catch (UnauthorizedHttpException $exception) {
-            $this->assertSame('Invalid two-factor code.', $exception->getMessage());
-            $this->assertSame(0, $pending->getFailedAttempts());
-        }
-    }
-
-    public function testInvokeThrowsUnauthorizedWhenPendingSessionConsumeFails(): void
-    {
-        $user = $this->createTwoFactorEnabledUser();
-        $pending = $this->createPendingSession($user->getId(), '+5 minutes');
-        $this->configureLookupsOnce($pending, $user);
-        $this->expectTotpVerification($user, '123456');
-        $this->pendingTwoFactorRepository->expects($this->once())
-            ->method('consumeIfActive')
-            ->with($pending->getId(), $this->isInstanceOf(DateTimeImmutable::class))
-            ->willReturn(false);
-        $this->sessionIssuer->expects($this->never())->method('create');
-        $this->events->expects($this->never())->method('publishCompleted');
-        $this->expectException(UnauthorizedHttpException::class);
-        $this->expectExceptionMessage('Invalid or expired two-factor session.');
-        $this->createHandler()->__invoke($this->createCommand($pending->getId(), '123456'));
-    }
-
-    private function assertInvalidTwoFactorCodeRejected(string $twoFactorCode): void
-    {
-        $user = $this->createTwoFactorEnabledUser();
-        $pending = $this->createPendingSession($user->getId(), '+5 minutes');
-        $this->configureLookupsOnce($pending, $user);
-        $this->twoFactorCodeVerifier->method('verifyAndResolveMethod')->willReturn(null);
-        $this->events->method('publishFailed');
-        try {
-            $this->createHandler()->__invoke(new CompleteTwoFactorCommand(
-                $pending->getId(),
-                $twoFactorCode,
-                $this->faker->ipv4(),
-                $this->faker->userAgent()
-            ));
-            $this->fail('Expected UnauthorizedHttpException to be thrown.');
-        } catch (UnauthorizedHttpException $exception) {
-            $this->assertSame('Invalid two-factor code.', $exception->getMessage());
-        }
-    }
-
-    private function arrangeRecoveryConsumeFailureAfterSessionConsumed(
-        PendingTwoFactor $pending,
-        User $user
-    ): void {
-        $this->configureLookupsOnce($pending, $user);
-        $this->twoFactorCodeVerifier->expects($this->once())
-            ->method('verifyAndResolveMethod')
-            ->with($user, 'AB12-CD34')
-            ->willReturn(TwoFactorCodeValidatorInterface::METHOD_RECOVERY_CODE);
-        $this->pendingTwoFactorRepository->expects($this->once())
-            ->method('consumeIfActive')
-            ->with($pending->getId(), $this->isInstanceOf(DateTimeImmutable::class))
-            ->willReturn(true);
-        $this->twoFactorCodeVerifier->expects($this->once())
-            ->method('consumeRecoveryCodeOrFail')
-            ->with($user, 'AB12-CD34')
-            ->willThrowException(
-                new UnauthorizedHttpException('Bearer', 'Invalid two-factor code.')
-            );
-        $this->expectNoBruteForceCounterAdvance($pending);
-    }
-
-    private function expectNoBruteForceCounterAdvance(PendingTwoFactor $pending): void
-    {
-        // The brute-force counter must NOT advance once the session is consumed.
-        $this->pendingTwoFactorRepository->expects($this->never())->method('save');
-        $this->pendingTwoFactorRepository->expects($this->never())->method('delete');
-        $this->events->expects($this->once())
-            ->method('publishFailed')
-            ->with($pending->getId(), $this->isType('string'), 'invalid_code');
-        $this->events->expects($this->never())->method('publishCompleted');
-    }
-
-    private function configureLookupsOnce(PendingTwoFactor $pending, User $user): void
-    {
-        $this->pendingTwoFactorRepository
-            ->expects($this->once())
-            ->method('findById')
-            ->with($pending->getId())
-            ->willReturn($pending);
-        $this->userRepository
-            ->expects($this->once())
-            ->method('findById')
-            ->with($user->getId())
-            ->willReturn($user);
-    }
-
-    private function expectTotpVerification(User $user, string $code): void
-    {
-        $this->twoFactorCodeVerifier->expects($this->once())
-            ->method('verifyAndResolveMethod')
-            ->with($user, $code)
-            ->willReturn('totp');
-    }
-
-    private function expectIssuedSession(IssuedSession $issued): void
-    {
-        $this->sessionIssuer->expects($this->once())
-            ->method('create')
-            ->willReturn($issued);
-    }
-
-    private function assertResponseTokens(
-        CompleteTwoFactorCommandResponse $response,
-        string $expectedAccessToken,
-        string $expectedRefreshToken
-    ): void {
-        $this->assertSame($expectedAccessToken, $response->getAccessToken());
-        $this->assertSame($expectedRefreshToken, $response->getRefreshToken());
-    }
-
-    private function createCommand(string $pendingId, string $code): CompleteTwoFactorCommand
-    {
-        return new CompleteTwoFactorCommand(
-            $pendingId,
-            $code,
-            $this->faker->ipv4(),
-            $this->faker->userAgent()
-        );
-    }
-
-    private function createHandler(): CompleteTwoFactorCommandHandler
-    {
-        return new CompleteTwoFactorCommandHandler(
-            $this->userRepository,
-            $this->pendingTwoFactorRepository,
-            $this->sessionIssuer,
-            $this->twoFactorCodeVerifier,
-            $this->events,
-        );
-    }
-
-    private function createUser(): User
-    {
-        return $this->userFactory->create(
-            $this->faker->email(),
-            $this->faker->firstName(),
-            $this->faker->password(),
-            $this->uuidTransformer->transformFromString($this->faker->uuid())
-        );
-    }
-
-    private function createTwoFactorEnabledUser(): User
-    {
-        $user = $this->createUser();
-        $user->setTwoFactorEnabled(true);
-        $user->setTwoFactorSecret('JBSWY3DPEHPK3PXP');
-        return $user;
-    }
-
-    private function createTwoFactorEnabledUserWithoutSecret(): User
-    {
-        $user = $this->createUser();
-        $user->setTwoFactorEnabled(true);
-        $user->setTwoFactorSecret(null);
-        return $user;
-    }
-
-    private function createPendingSession(
-        string $userId,
-        string $expiresAtModifier
-    ): PendingTwoFactor {
-        $createdAt = new DateTimeImmutable('now');
-
-        return new PendingTwoFactor(
-            (string) new Ulid(),
-            $userId,
-            $createdAt,
-            $createdAt->modify($expiresAtModifier)
-        );
     }
 }
